Rename DGUS files
Isolate the rename part of #17142.
