+++ Marlin/src/lcd/extensible_ui/lib/dgus/DGUSDisplay.cpp
@@ -32,7 +32,7 @@
 
 #include "DGUSDisplay.h"
 #include "DGUSVPVariable.h"
-#include "DGUSDisplayDefinition.h"
+#include "DGUSDisplayDef.h"
 
 #include "../../ui_api.h"
